update moved dependencies
