+++ client/src/pages/lab_management/users/PrivilegeControl.tsx
@@ -7,10 +7,10 @@ import { FullMakerspace, GET_FULL_MAKERSPACES } from "../../../queries/makerspac
 import RequestWrapper2 from "../../../common/RequestWrapper2";
 import { isManagerFor } from "../../../common/PrivilegeUtils";
 import DeleteIcon from '@mui/icons-material/Delete';
-import { GET_USER } from "./UserModal";
 import TrainerCard from "./TrainerCard";
 import TrainerEquipmentSelect from "./TrainerEquipmentSelect";
 import PersonAddIcon from '@mui/icons-material/PersonAdd';
+import { GET_USER } from "../../../queries/userQueries";
 
 
 interface PrivilegeControlProps {
